Rename worker that only fetches meals for one day
