+++ app/src/main/kotlin/de/janhopp/luebeckmensawidget/ui/activity/MensaDayActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import de.janhopp.luebeckmensawidget.SyncWorker.Companion.enqueueSyncWork
+import de.janhopp.luebeckmensawidget.DaySyncWorker.Companion.enqueueSyncWork
 import de.janhopp.luebeckmensawidget.ui.theme.MensaTheme
 
 class MensaDayActivity: ComponentActivity() {
